# api-spec.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API Pelacakan Alumni STIS",
        "description": "REST API untuk mengelola profil alumni, riwayat pekerjaan, serta statistik terkait.",
        "contact": {
            "name": "Pelacakan Alumni STIS",
            "email": "admin@stis.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Registrasi pengguna baru",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAuthRegisterResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Masuk dan mendapatkan token JWT",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAuthLoginResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/dashboard/summary": {
            "get": {
                "tags": [
                    "Dasbor"
                ],
                "summary": "Mengambil ringkasan metrik dasbor",
                "operationId": "summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseDashboardSummaryResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthRegisterRequest": {
                "required": [
                    "angkatan",
                    "confirmPassword",
                    "email",
                    "fullName",
                    "nim",
                    "password",
                    "programStudi",
                    "tahunLulus",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "maxLength": 50,
                        "minLength": 4,
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 100,
                        "minLength": 8,
                        "type": "string"
                    },
                    "confirmPassword": {
                        "maxLength": 100,
                        "minLength": 8,
                        "type": "string"
                    },
                    "fullName": {
                        "maxLength": 150,
                        "minLength": 0,
                        "type": "string"
                    },
                    "nim": {
                        "maxLength": 20,
                        "minLength": 0,
                        "type": "string"
                    },
                    "angkatan": {
                        "maximum": 67,
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    },
                    "programStudi": {
                        "pattern": "(?i)^(D1|D3|D4)$",
                        "type": "string"
                    },
                    "tahunLulus": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "phoneNumber": {
                        "maxLength": 30,
                        "minLength": 0,
                        "type": "string"
                    },
                    "alamat": {
                        "maxLength": 255,
                        "minLength": 0,
                        "type": "string"
                    }
                }
            },
            "ApiResponseAuthRegisterResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AuthRegisterResponse"
                    }
                }
            },
            "AuthRegisterResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "nim": {
                        "type": "string"
                    },
                    "angkatan": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "programStudi": {
                        "type": "string"
                    },
                    "tahunLulus": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "AuthLoginRequest": {
                "required": [
                    "password",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "ApiResponseAuthLoginResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AuthLoginResponse"
                    }
                }
            },
            "AuthLoginResponse": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string"
                    },
                    "tokenType": {
                        "type": "string"
                    },
                    "expiresIn": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserSummaryResponse"
                    }
                }
            },
            "UserSummaryResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "USER",
                            "ADMIN"
                        ]
                    }
                }
            },
...
            "ApiResponseDashboardSummaryResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/DashboardSummaryResponse"
                    }
                }
            },
            "DashboardSummaryResponse": {
                "type": "object",
                "properties": {
                    "totalAlumni": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "activeAlumni": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "inactiveAlumni": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "workingAlumni": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nonWorkingAlumni": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalInstitutions": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}